backend/requirements.txt
opencv-python
numpy
flask
flask-socketio
python-socketio